test: refactor test fixtures
