+++ traveller/conftest.py
@@ -13,13 +13,11 @@ from flask import url_for
 
 from app import create_app
 from init import db as _db
-from modules.box__default.auth.models import User, Role
+from modules.box__default.auth.models import User
 from modules.box__default.settings.models import Settings
 
 # run in shopyo/shopyo
 # python -m pytest . or python -m pytest -v
-from modules.conf.models import Conf, Talk, AuthorList, ReviewerList
-from modules.schedule.models import Schedule, Day, Activity
 
 if os.path.exists("testing.db"):
     os.remove("testing.db")
@@ -67,84 +65,6 @@ def admin_user():
 
 
 @pytest.fixture(scope="session")
-def conf(current_year):
-    """
-    A pytest fixture that returns a conference for the current year
-    """
-
-    now = datetime.datetime.now()
-    _30_days_ago = now - datetime.timedelta(days=30)
-    _30_days_from_now = now + datetime.timedelta(days=30)
-    conf = Conf(id=1)
-    conf.year = current_year
-    conf.cfp_start = _30_days_ago.date()
-    conf.cfp_end = _30_days_from_now.date()
-    conf.reviewer_list = ReviewerList()
-    return conf
-
-
-@pytest.fixture(scope="session")
-def last_conf(current_year):
-    """
-    A pytest fixture that returns a conference for the last year
-    """
-
-    now = datetime.datetime.now()
-    _30_days_ago = now - datetime.timedelta(days=365+30)
-    _30_days_from_now = now - datetime.timedelta(days=365-30)
-    conf = Conf()
-    conf.year = current_year - 1
-    conf.cfp_start = _30_days_ago.date()
-    conf.cfp_end = _30_days_from_now.date()
-    return conf
-
-
-@pytest.fixture(scope="session")
-def talk(current_year):
-    return Talk(
-        id=1,
-        title="Test Talk",
-        summary="Some summary",
-        description="Some Description",
-        level="beginner",
-        year=current_year
-    )
-
-
-@pytest.fixture(scope="session")
-def last_talk(current_year):
-    return Talk(
-        id=2,
-        title="Test Talk2",
-        summary="Some summary",
-        description="Some Description",
-        level="beginner",
-        year=current_year - 1
-    )
-
-
-@pytest.fixture(scope="session")
-def schedule(conf):
-    return Schedule(
-        days=[
-            Day(
-                id=1,
-                date=conf.cfp_end,
-                activities=[
-                    Activity(
-                        type="normal_activity",
-                        text="Fun and Games",
-                        start_time=datetime.time(9, 30),
-                        end_time=datetime.time(10, 0),
-                    )
-                ],
-            ),
-        ],
-        conf_id=conf.id
-    )
-
-
-@pytest.fixture(scope="session")
 def flask_app():
     flask_app = create_app("testing")
     return flask_app
@@ -163,7 +83,7 @@ def test_client(flask_app):
 
 
 @pytest.fixture(scope="session")
-def db(test_client, non_admin_user, admin_user, unconfirmed_user, conf, last_conf, talk, last_talk, schedule):
+def db(test_client, non_admin_user, admin_user, unconfirmed_user):
     """
     creates and returns the initial testing database
     """
@@ -175,25 +95,7 @@ def db(test_client, non_admin_user, admin_user, unconfirmed_user, conf, last_con
     _db.session.add(non_admin_user)
     _db.session.add(admin_user)
     _db.session.add(unconfirmed_user)
-    conf.reviewer_list.reviewers.append(non_admin_user)
-    _db.session.add(conf)
-    _db.session.add(last_conf)
 
-    talk.create_slug()
-    talk.author_list = AuthorList()
-    talk.author_list.authors.append(non_admin_user)
-    talk.submitter_id = non_admin_user.id
-    talk.talk_conference = conf
-    _db.session.add(talk)
-
-    last_talk.create_slug()
-    last_talk.author_list = AuthorList()
-    last_talk.author_list.authors.append(non_admin_user)
-    last_talk.submitter_id = non_admin_user.id
-    last_talk.talk_conference = last_conf
-    _db.session.add(last_talk)
-
-    _db.session.add(schedule)
 
     config_json_file = pathlib.Path(__file__).parent.resolve().joinpath("config.json")
     # add the default settings
